Set smoothing in cube_to_target via ifdef
Generating topography for new grids requires compiling and running
cube_to_target twice: once with the smoothing flags on, and once with
them off. Previously, this required editing the source code for
cube_to_target and recompiling. Instead, so that we do not need to edit
the source code (and hence can script this process), we set the flags
via preprocessor directives, which we can set via commands to make when
building. This is not the best solution (should enable arguments to be
passed to set these dynamically probably), but will work fine for now.