Fix unallocated phisdyn error
Topography fields (i.e. PHISDYN) were being deallocated before their
last use when using the topography smoothing subroutines in the SE
dynamical codes. The temporary fix was to remove the deallocate
statements from the initialization routine, but now we need to figure
out where it is safe to deallocate these.