Use 30 ninst for MVK CIME System Test
The MVK test was developed to use 30 instances initially, but as
described in issue #2487, 30 instances would cause the MPI tags to
overflow. However, reverting PR #2231 reduces the max MPI tag max value
and again allows for 30 instances to be run.

Now, we can use 30 instances and not have to bootstrap the critical
value again.