Fix optical depth for MODIS simulator
Fix setting of optical depth in MODIS simulator. There was a typo in
that the convective optical depth was being used instead of the snow
optical depth. This is expected to affect all of the modis diagnostics
in cases in which snow is present, but should not affect the model
solution since this is purely diagnostic.