Fix Makefile for cube_to_target topo tool
Fix Makefile to build the cube_to_target tool that interpolates USGS
high resolution topography to a target grid. This Makefile only included
flags to build with the lf95 compiler. This will now build using ifort
if the FC environment variable is set to ifort. NetCDF library paths are
pulled from the LIB_NETCDF and INC_NETCDF environment variales, so if
these are not set and the libraries are in non-standard locations this
will probably fail. Sourcing an env_mach_specific.sh from a working E3SM
case should set these appropriately. This builds without issue on SNL
sandiatoss3 machines (tested on Chama).